requirements.txt
torch
torchvision
opencv-python
Pillow
reportlab